folder restructuring + UpdateDraftRelationships command wip
