+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Contexts/ValidateAndRepairItemSyncContext.cs
@@ -1,5 +1,5 @@
 ﻿using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces;
-using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Models;
+using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces;
 using DFC.ServiceTaxonomy.GraphSync.Services.Interface;
 using OrchardCore.ContentManagement;
 using OrchardCore.ContentManagement.Metadata.Models;
