+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Helpers/GraphValidationHelper.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces;
-using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Models;
+using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces;
 using DFC.ServiceTaxonomy.Neo4j.Extensions;
 using Neo4j.Driver;
 using Newtonsoft.Json.Linq;
