+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Interfaces/IGraphValidationHelper.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Models;
+using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces;
 using Neo4j.Driver;
 using Newtonsoft.Json.Linq;
