+++ DFC.ServiceTaxonomy.GraphSync/Neo4j/Commands/Interfaces/IUpdateDraftRelationships.cs
@@ -1,0 +1,9 @@
+﻿using DFC.ServiceTaxonomy.Neo4j.Commands.Interfaces;
+
+namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Commands.Interfaces
+{
+    public interface IUpdateDraftRelationships : ICommand
+    {
+
+    }
+}
