+++ DFC.ServiceTaxonomy.GraphSync/Neo4j/Commands/UpdateDraftRelationships.cs
@@ -1,0 +1,7 @@
+﻿namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Commands
+{
+    public class UpdateDraftRelationships
+    {
+        
+    }
+}
