+++ DFC.ServiceTaxonomy.GraphSync/Neo4j/Queries/Interfaces/IGetContentItemsAsJsonQuery.cs
@@ -1,6 +1,6 @@
 ﻿using DFC.ServiceTaxonomy.Neo4j.Queries.Interfaces;
 
-namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries
+namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces
 {
     interface IGetContentItemsAsJsonQuery : IQuery<string>
     {
