+++ DFC.ServiceTaxonomy.GraphSync/Neo4j/Queries/Interfaces/INodeWithOutgoingRelationships.cs
@@ -3,7 +3,7 @@ using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces;
 using DFC.ServiceTaxonomy.Neo4j.Commands.Model;
 using Neo4j.Driver;
 
-namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Models
+namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces
 {
     public interface INodeWithOutgoingRelationships
     {
