+++ DFC.ServiceTaxonomy.GraphSync/Neo4j/Queries/Interfaces/IOutgoingRelationship.cs
@@ -1,6 +1,6 @@
 ﻿using Neo4j.Driver;
 
-namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Models
+namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces
 {
     public interface IOutgoingRelationship
     {
