+++ DFC.ServiceTaxonomy.GraphSync/Neo4j/Queries/Models/NodeWithOutgoingRelationships.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces;
+using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces;
 using DFC.ServiceTaxonomy.Neo4j.Commands.Model;
 using Neo4j.Driver;
 using OrchardCore.Workflows.Helpers;
