+++ DFC.ServiceTaxonomy.GraphSync/Neo4j/Queries/Models/OutgoingRelationship.cs
@@ -1,4 +1,5 @@
-﻿using Neo4j.Driver;
+﻿using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces;
+using Neo4j.Driver;
 
 namespace DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Models
 {
