+++ DFC.ServiceTaxonomy.UnitTests/UnitTestHelpers/PartGraphSyncer/PartGraphSyncer_ValidateSyncComponentTests.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces;
-using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Models;
+using DFC.ServiceTaxonomy.GraphSync.Neo4j.Queries.Interfaces;
 using DFC.ServiceTaxonomy.GraphSync.Services.Interface;
 using FakeItEasy;
 using Neo4j.Driver;
